+++ src/app/users.js
@@ -1,11 +1,37 @@
 import express from 'express';
 import User from '../models/user.js';
+import tokenChecker from './tokenChecker.js';
 
 const router = express.Router();
 
-//aggiungere get /me
+router.get('/me', tokenChecker, async function(req,res){
+    if(!req.loggedUser){
+        res.status(401).json({error: 'You are not authenticated'});
+        return;
+    }
 
-router.get('', async function(req,res){
+    let user= await User.findOne({email: req.loggedUser.email}).exec();
+
+    user= {
+            self: 'renTrentoAPI/users' + user.id,
+            userName: user.userName,
+            email: user.email,
+            wallet: user.wallet
+        }
+    
+    res.status(200).json(user);
+})
+
+router.get('', tokenChecker, async function(req,res){
+    if(!req.loggedUser){
+        res.status(401).json({error: 'You are not authenticated'});
+        return;
+    }
+    if(req.loggedUser.role!='admin'){
+        res.status(403).json({error: 'Yuo are not allowed to do this' })
+        return;
+    }
+    
     let users = await User.find().exec();
 
     users= users.map(function(user){
@@ -34,22 +60,9 @@ router.post('', async function(req,res){
     }
 
     user = await user.save();
-    
-    //let userId = user._id;
 
     res.status(201).json(user);
 })
-
-/*router.use('/:userId', async function(req,res,next){
-    let user= await User.findById(req.params.userId).exec();
-    if(!user){
-        res.status(404).send()
-        console.log('user not found')
-        return;
-    }
-    req['user'] = user;
-    next();
-})*/
 
 router.get('/:userId', async function(req,res){
     let user= await User.findById(req.params.userId).exec();
@@ -67,25 +80,49 @@ router.get('/:userId', async function(req,res){
     res.status(200).json(user);
 })
 
-router.delete('/:userId', async function(req,res){
+router.delete('/:userId', tokenChecker, async function(req,res){
+    if(!req.loggedUser){
+        res.status(401).json({error: 'You are not authenticated'});
+        return;
+    }
+
     let user= await User.findById(req.params.userId).exec();
+
+    if(!user){
+        res.status(404).send();
+        console.log('user not found');
+        return;
+    }
+    
+    if(req.loggedUser.role!='admin' && req.loggedUser.id!=user.id){
+        res.status(403).json({ error: 'Yuo are not allowed to do this' })
+        return;
+    }
+
+    await User.deleteOne({ _id: user.id});
+    console.log('account deleted')
+    res.status(204).send();
+})
+
+router.patch('/:userId', tokenChecker, async function(req,res){
+    if(!req.loggedUser){
+        res.status(401).json({error: 'You are not authenticated'});
+        return;
+    }
+
+    let user= await User.findById(req.params.userId).exec();
+
     if(!user){
         res.status(404).send()
         console.log('user not found')
         return;
     }
-    await User.deleteOne({ _id: user.id});
-    console.log('account deleted')
-    res.status(204).send()
-})
-
-router.patch('/:userId', async function(req,res){
-    let user= await User.findById(req.params.userId).exec();
-    /*if(!user){
-        res.status(404).send()
-        console.log('user not found')
+    
+    if(req.loggedUser.role!='admin' && req.loggedUser.id!=user.id){
+        res.status(403).json({ error: 'Yuo are not allowed to do this' })
         return;
-    }*/
+    }
+    
     if(req.body.userName)
         user.userName= req.body.userName;
     if(req.body.role)
